Move connection into show
